Create Indexer
This is what I have so far. I tried to keep the objects added to the database in the expected structure but I will walk thru my entire code again soon so that it is readable and does exactly what is expected. I will probably break down the methods into smaller, more efficient ones.